move formplayer deploy code to its own module
