+++ src/commcare_cloud/commands/deploy.py
@@ -14,6 +14,7 @@ from commcare_cloud.commands.ansible import ansible_playbook
 from commcare_cloud.commands.ansible.helpers import AnsibleContext
 from commcare_cloud.commands.command_base import Argument, CommandBase
 from commcare_cloud.commands.deploy.commcare import deploy_commcare
+from commcare_cloud.commands.deploy.formplayer import deploy_formplayer
 from commcare_cloud.commands.terraform.aws import get_default_username
 from commcare_cloud.environment.main import get_environment
 from commcare_cloud.environment.paths import get_available_envs
@@ -74,52 +75,14 @@ class Deploy(CommandBase):
         if deploy_component in ['commcare', 'both']:
             if deploy_component != 'both':
                 _warn_no_formplayer()
-            deploy_commcare(environment, args, unknown_args)
+            return deploy_commcare(environment, args, unknown_args)
         if deploy_component in ['formplayer', 'both']:
             if deploy_component != 'both':
                 if args.commcare_rev:
                     print(color_warning('--commcare-rev does not apply to a formplayer deploy and will be ignored'))
                 if args.fab_settings:
                     print(color_warning('--set does not apply to a formplayer deploy and will be ignored'))
-            self._announce_formplayer_deploy_start(environment)
-            self.deploy_formplayer(environment, args, unknown_args)
-
-    def deploy_formplayer(self, environment, args, unknown_args):
-        def run_ansible_playbook_command():
-            skip_check = True
-            environment.create_generated_yml()
-            ansible_context = AnsibleContext(args)
-            return ansible_playbook.run_ansible_playbook(
-                environment, 'deploy_stack.yml', ansible_context,
-                skip_check=skip_check, quiet=skip_check, always_skip_check=skip_check, limit='formplayer',
-                use_factory_auth=False, unknown_args=('--tags=formplayer_deploy',),
-                respect_ansible_skip=True,
-            )
-        rc = run_ansible_playbook_command()
-        if rc != 0:
-            return rc
-        rc = commcare_cloud(
-            args.env_name, 'run-shell-command', 'formplayer',
-            ('supervisorctl reread; '
-             'supervisorctl update {project}-{deploy_env}-formsplayer-spring; '
-             'supervisorctl restart {project}-{deploy_env}-formsplayer-spring').format(
-                project='commcare-hq',
-                deploy_env=environment.meta_config.deploy_env,
-            ), '-b',
-        )
-        if rc != 0:
-            return rc
-
-    @staticmethod
-    def _announce_formplayer_deploy_start(environment):
-        mail_admins(
-            environment,
-            subject="{user} has initiated a formplayer deploy to {environment}.".format(
-                user=get_default_username(),
-                environment=environment.meta_config.deploy_env,
-            ),
-            message='',
-        )
+            return deploy_formplayer(environment, args)
 
 
 def _warn_no_formplayer():
@@ -130,13 +93,3 @@ def _warn_no_formplayer():
         Preview, but keeps these services up to date.
         You can do so by running `commcare-cloud <env> deploy formplayer`
     """)))
-
-
-def mail_admins(environment, subject, message):
-    if environment.fab_settings_config.email_enabled:
-        commcare_cloud(
-            environment.name, 'django-manage', 'mail_admins',
-            '--subject', subject,
-            message,
-            '--environment', environment.meta_config.deploy_env
-        )
